Do not load configuration globally
Make commands responsible for loading configuration.
