+++ src/gsebuild/config.ts
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import fs from "node:fs/promises";
+import path from "node:path";
+
 /**
  * A pattern to copy.
  *
@@ -169,3 +172,47 @@ export interface Configuration {
 export type ExpandedConfiguration = Required<{
   readonly [P in keyof Configuration]: Required<Configuration[P]>;
 }>;
+
+interface PackageJson {
+  readonly gsebuild?: Configuration;
+}
+
+/**
+ * Get the expanded configuration.
+ *
+ * Load configuration from `package.json` and expand missing values to their
+ * defaults.
+ *
+ * @returns The configuration from `package.json`
+ */
+export const getConfiguration = async (): Promise<ExpandedConfiguration> => {
+  const packageDirectory = path.resolve();
+  const config = (
+    JSON.parse(
+      await fs.readFile(path.join(packageDirectory, "package.json"), {
+        encoding: "utf-8",
+      }),
+    ) as PackageJson
+  ).gsebuild;
+  return {
+    extension: {
+      "metadata-file":
+        config?.extension?.["metadata-file"] ??
+        path.join(packageDirectory, "metadata.json"),
+      "po-directory":
+        config?.extension?.["po-directory"] ??
+        path.join(packageDirectory, "po"),
+    },
+    pack: {
+      "copy-to-source": config?.pack?.["copy-to-source"] ?? [],
+      "extra-sources": config?.pack?.["extra-sources"] ?? [],
+      schemas: config?.pack?.schemas ?? [],
+      "source-directory":
+        config?.pack?.["source-directory"] ?? packageDirectory,
+    },
+    gettext: {
+      sources: config?.gettext?.sources ?? [],
+      "copyright-holder": config?.gettext?.["copyright-holder"] ?? null,
+    },
+  };
+};
